refactor: HomeModule CQRS 패턴 시범적용
