+++ src/app.module.ts
@@ -93,7 +93,8 @@ export class AppModule implements OnModuleInit, NestModule {
       await this.seederService.initializeServer();
       await this.seederService.initializeAdmin();
       await this.seederService.initializeAppVersions();
-      // await this.seederService.initializeNicknames();
+      await this.seederService.initializeDefaultTheme();
+      // await this.seederService.initializeNicknames()
     }
 
     this.configService.set('APP_INITIALIZING', false);
